Rename file key-numbers.vue to KeyNumbers.vue
