+++ pages/higher-education.vue
@@ -87,7 +87,7 @@
 import { documents, documentFetcher } from '~/services/document-fetcher'
 import SectionSlice from '@/components/slices/Section'
 import SectionColumnSlice from '@/components/slices/SectionColumn'
-import KeyNumbers from '~/components/key-numbers'
+import KeyNumbers from '@/components/KeyNumbers'
 
 export default {
   name: 'HigherEducation',
